@@ -1,5 +1,8 @@
 import { Component } from '@angular/core';
-import { NavController } from 'ionic-angular';
+import { NavController, NavParams } from 'ionic-angular';
+
+import { AddressPersonDetailPage } from './../../pages/pages';
+import { SubjectProvider, Subject, User } from './../../providers/api-services/subjects';
 
 @Component({
   selector: 'component-subject-student',
@@ -7,7 +10,37 @@ import { NavController } from 'ionic-angular';
 })
 export class SubjectStudentCompnent {
 
-  constructor(public navCtrl: NavController) {
+  public subject: Subject;
+  public users: User[];
+  public loading: boolean = false;
+
+  constructor(
+    private navCtrl: NavController,
+    private navParams: NavParams,
+    private subjectProvider: SubjectProvider,
+  ) {
+    this.subject = this.navParams.data;
+  }
+
+  public ngOnInit(): void {
+    this.loadUsers();
+  }
+
+  public goToDetail(user: User): void {
+    this.navCtrl.push(AddressPersonDetailPage, user);
+  }
+
+  private loadUsers(): void {
+    if (this.subject.id) {
+      this.users = [];
+      this.loading = true;
+      this.subjectProvider.getUsers(this.subject.id)
+        .subscribe(data => {
+          this.loading = false;
+          this.users = data;
+        });
+
+    }
   }
 
 }
