@@ -1,13 +1,40 @@
 import { Component } from '@angular/core';
 import { NavController, NavParams } from 'ionic-angular';
 
+import { TaskProvider, Task, TaskItem } from './../../providers/api-services/tasks';
+
 @Component({
   selector: 'compnent-task-item',
   templateUrl: 'task-item.html',
 })
 export class TaskItemCompnent {
 
-  constructor(public navCtrl: NavController, public navParams: NavParams) {
+  public task: Task;
+  public taskItems: TaskItem[];
+  public loading: boolean = false;
+
+  constructor(
+    private navCtrl: NavController,
+    private navParams: NavParams,
+    private taskProvider: TaskProvider,
+  ) {
+    this.task = this.navParams.data;
+  }
+
+  public ngOnInit(): void {
+    if (this.task.id) {
+      this.loadTaskItems();
+    }
+  }
+
+  private loadTaskItems(): void {
+    this.loading = true;
+    this.taskItems = [];
+    this.taskProvider.getTaskItems(this.task.id)
+      .subscribe(data => {
+        this.loading = false;
+        this.taskItems = data;
+      })
   }
 
 }
