@@ -7,6 +7,6 @@ export const devVariables = {
     clientSecret: 'ME0_PXJazxvf4xZo77OzryyVrKcMGQ_1csbWC8gJER1fcvIqHQ92WZr_6FReAMLP',
     domain: 'dipl.eu.auth0.com',
     audience: 'http://127.0.0.1:8000',
-    scope: 'openid'
+    scope: 'openid email profile'
   }
 };
