@@ -1,14 +1,21 @@
-// The page the user lands on after opening the app and without a session
-export const FirstRunPage = 'TutorialPage';
-
 // The main page the user will see as they use the app over a long period of time.
 // Change this if not using tabs
 export const MainPage = 'DashboardPage';
+
+// Tutorial
+export const TutorialPage = 'TutorialPage';
+
+// Dashboard
+export const DashboardPage = 'DashboardPage';
+
+// Grade
+export const GradePage = 'GradePage';
 
 // Login
 export const LoginPage = 'LoginPage';
 
 // Chat pages
+export const ChatPage = 'ChatPage';
 export const ChatDetailPage = 'ChatDetailPage';
 
 // Notifications
@@ -20,6 +27,7 @@ export const AddressPersonDetailPage = 'AddressPersonDetailPage';
 export const AddressGroupDetailPage = 'AddressGroupDetailPage';
 
 // Subject
+export const SubjectPage = 'SubjectPage';
 export const SubjectDetailPage = 'SubjectDetailPage';
 
 // Lesson
@@ -27,3 +35,6 @@ export const LessonDetailPage = 'LessonDetailPage';
 
 // Task
 export const TaskDetailPage = 'TaskDetailPage';
+
+// Settings
+export const SettingsPage = 'SettingsPage';
